view fixed and other stuff from last pull request (#59)
* lanecomponent basis

* component id for lanes

* lanecomponent finished basis

* filters

* lanecomp gone, Bart will make this

* lane object

* mtlfile added and lane_component enum deleted

* new lane object

* object with texture

* lanepart texture

* lanepart.Cobj width length hight corrected

* filters and score commented

* scale component, added skybox with scale, moved camera to right position above lanes

* removed commentlines

* camera rotation

* working meshdraw methods

* transporter model and pullrequest improvements

* model

* models updated and alphablending testing

* lanecolor

* heart object

* model.cpp repair

* terminatethead!!

* fixed viewproblems

* lanedist smaller

* initsound first

+++ ProftaakPeriode4/ProftaakPeriode4/VisionComponent.h
@@ -92,6 +92,6 @@ private:
 	cv::Mat _src_gray;
 	const int _thresh = 50;
 	const int max_thresh = 255;
-	const int _radius = 10;
+	const float _radius = 10;
 };
